Fix reversal bot creation
